refactoring for prelude use
