+++ cw-orch/src/contract.rs
@@ -1,8 +1,8 @@
-use crate::tx_handler::ChainUpload;
+use crate::environment::ChainUpload;
+use crate::prelude::{CwEnv, Uploadable};
 use crate::{
-    error::CwOrcError, index_response::IndexResponse, state::StateInterface, tx_handler::TxResponse,
+    error::CwOrcError, index_response::IndexResponse, state::StateInterface, environment::TxResponse,
 };
-use crate::{CwEnv, Uploadable};
 use cosmwasm_std::{Addr, Coin};
 use serde::{de::DeserializeOwned, Serialize};
 use std::fmt::Debug;
