+++ cw-orch/src/daemon/traits.rs
@@ -1,4 +1,7 @@
-use crate::{queriers::CosmWasm, CwOrcError, CwOrcMigrate, CwOrcUpload, Daemon, TxResponse};
+use crate::environment::TxResponse;
+use crate::error::CwOrcError;
+use crate::prelude::*;
+use crate::daemon::queriers::CosmWasm;
 
 pub trait UploadHelpers: CwOrcUpload<Daemon> {
     /// Only upload the contract if it is not uploaded yet (checksum does not match)
